+++ utils/get-tooling.js
@@ -7,7 +7,7 @@ module.exports = options => {
   // add backdrush
   if (options.backdrush !== false) tooling.drush = {service: 'appserver'};
   // add bee
-  if (options.bee !== false) tooling.bee = {service: 'appserver', cmd: `/var/www/.bee/bee-${options.bee}/bee.php`};
+  if (options.bee !== false) tooling.bee = {service: 'appserver'};
 
   return tooling;
 };